Enforce strict permission checks for company and employee actions, update UI for permission-based buttons

+++ app/Http/Controllers/CompanyController.php
@@ -4,22 +4,36 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Company;
+use App\Helpers\PermissionHelper;
 
 class CompanyController extends Controller
 {
     public function index()
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'view company')) {
+            abort(403, 'You do not have permission to view companies.');
+        }
         $companies = Company::all();
         return view('companies.index', compact('companies'));
     }
 
     public function create()
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'create company')) {
+            abort(403, 'You do not have permission to create companies.');
+        }
         return view('companies.create');
     }
 
     public function store(Request $request)
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'create company')) {
+            abort(403, 'You do not have permission to create companies.');
+        }
+
         $request->validate([
             'company_name' => 'required|string|max:255',
             'address' => 'required|string',
@@ -41,16 +55,29 @@ class CompanyController extends Controller
 
     public function show(Company $company)
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'view company')) {
+            abort(403, 'You do not have permission to view companies.');
+        }
         return view('companies.show', compact('company'));
     }
 
     public function edit(Company $company)
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'edit company')) {
+            abort(403, 'You do not have permission to edit companies.');
+        }
         return view('companies.edit', compact('company'));
     }
 
     public function update(Request $request, Company $company)
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'edit company')) {
+            abort(403, 'You do not have permission to edit companies.');
+        }
+
         $request->validate([
             'company_name' => 'required|string|max:255',
             'address' => 'required|string',
@@ -72,6 +99,10 @@ class CompanyController extends Controller
 
     public function destroy(Company $company)
     {
+        $user = auth()->user();
+        if (!PermissionHelper::isUserPermittedTo($user, 'delete company')) {
+            abort(403, 'You do not have permission to delete companies.');
+        }
         $company->delete();
         return redirect()->route('companies.index')->with('success', 'Company deleted successfully.');
     }
